.github/scripts/requirements.txt
# Dependencies untuk sitemap generation dan Google Search Console
lxml>=4.9.0
beautifulsoup4>=4.12.0
google-auth>=2.23.0
google-api-python-client>=2.108.0